Fix: gateway service
