+++ api-gateway/server.js
@@ -1,0 +1,73 @@
+import express from "express";
+import cors from "cors";
+import { createProxyMiddleware } from "http-proxy-middleware";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+const app = express();
+const PORT = process.env.PORT || 3000;
+
+// Middleware
+app.use(cors());
+app.use(express.json());
+
+// Service URLs from environment variables
+const PRODUCT_SERVICE_URL = process.env.PRODUCT_SERVICE_URL;
+const USER_SERVICE_URL = process.env.USER_SERVICE_URL;
+const CART_SERVICE_URL = process.env.CART_SERVICE_URL;
+const ORDER_SERVICE_URL = process.env.ORDER_SERVICE_URL;
+
+// Routes
+app.use(
+  "/api/products",
+  createProxyMiddleware({
+    target: PRODUCT_SERVICE_URL,
+    changeOrigin: true,
+    pathRewrite: {
+      "^/api/products": "/api/products",
+    },
+  })
+);
+
+app.use(
+  "/api/users",
+  createProxyMiddleware({
+    target: USER_SERVICE_URL,
+    changeOrigin: true,
+    pathRewrite: {
+      "^/api/users": "/api/users",
+    },
+  })
+);
+
+app.use(
+  "/api/cart",
+  createProxyMiddleware({
+    target: CART_SERVICE_URL,
+    changeOrigin: true,
+    pathRewrite: {
+      "^/api/cart": "/api/cart",
+    },
+  })
+);
+
+app.use(
+  "/api/orders",
+  createProxyMiddleware({
+    target: ORDER_SERVICE_URL,
+    changeOrigin: true,
+    pathRewrite: {
+      "^/api/orders": "/api/orders",
+    },
+  })
+);
+
+// Health check endpoint
+app.get("/health", (req, res) => {
+  res.status(200).json({ status: "API Gateway is running" });
+});
+
+app.listen(PORT, () => {
+  console.log(`API Gateway running on port ${PORT}`);
+});
