--- gateway/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "gateway",
-  "version": "1.0.0",
-  "description": "API Gateways",
-  "main": "server.js",
-  "scripts": {
-    "start": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0",
-    "helmet": "^8.1.0",
-    "http-proxy-middleware": "^3.0.5",
-    "mongoose": "^8.13.3",
-    "morgan": "^1.10.0",
-    "nodemon": "^3.1.10"
-  }
-}
